requirements.txt
altgraph==0.17.2
bitstring==3.1.9
future==0.18.2
khbr==4.0.5
pefile==2021.9.3
pillow==9.3.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyperclip==1.8.2
PySide6==6.2.1
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==5.4.1
qt-material==2.8.13
shiboken6==6.2.1
tomlkit==0.11.1